+++ packages/app/integration_test/rust_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:keychat_rust_ffi_plugin/index.dart';
@@ -8,7 +9,7 @@ void main() {
   setUpAll(() async => await RustLib.init());
   test('Can call rust function', () async {
     var a = await rust_nostr.generateSecp256K1();
-    print(a.prikey);
+    debugPrint(a.prikey);
     expect(a.pubkey.length, 64);
   });
 }