+++ packages/app/lib/page/setting/app_general_setting.dart
@@ -2,16 +2,20 @@
 
 import 'dart:io' show File, exit;
 
+import 'package:app/controller/setting.controller.dart';
 import 'package:app/global.dart';
+import 'package:app/models/db_provider.dart';
 import 'package:app/page/dbSetup/db_setting.dart';
 
 import 'package:app/controller/home.controller.dart';
 import 'package:app/page/FileExplore.dart';
 import 'package:app/page/login/OnboardingPage2.dart';
+import 'package:app/page/routes.dart';
 import 'package:app/page/widgets/notice_text_widget.dart';
 import 'package:app/service/file.service.dart';
 import 'package:app/service/notify.service.dart';
 import 'package:app/service/secure_storage.dart';
+import 'package:app/service/storage.dart';
 import 'package:app/service/websocket.service.dart';
 import 'package:app/utils.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
@@ -22,12 +26,6 @@ import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:get/get.dart';
 
 import 'package:settings_ui/settings_ui.dart';
-// import 'package:restart_app/restart_app.dart';
-
-import '../../controller/setting.controller.dart';
-import '../../models/db_provider.dart';
-import '../../service/storage.dart';
-import '../routes.dart';
 
 class AppGeneralSetting extends GetView<SettingController> {
   const AppGeneralSetting({super.key});